Temporarily disable event listeners.
When a user pushes vigorously, s/he often immediately pulls back. (Also vice
versa.) This can trigger a push event followed by a pull event. We want only
the push event to be triggered.

To prevent triggering the second event, we disable its event listener for a
brief period given by the variable 'delay'. This is done by setting a timeout
to restore the event listener and then removing it.